feat(agents): extract control plane APIs
