+++ services/agents/package.json
@@ -8,6 +8,12 @@
     "./server/controller-runtime": "./src/server/controller-runtime.ts",
     "./server/controller-runtime-config": "./src/server/controller-runtime-config.ts",
     "./server/control-plane": "./src/server/control-plane.ts",
+    "./server/control-plane-cache": "./src/server/control-plane-cache.ts",
+    "./server/control-plane-cache-config": "./src/server/control-plane-cache-config.ts",
+    "./server/control-plane-cache-freshness": "./src/server/control-plane-cache-freshness.ts",
+    "./server/control-plane-cache-store": "./src/server/control-plane-cache-store.ts",
+    "./server/control-plane-primitive-kinds": "./src/server/control-plane-primitive-kinds.ts",
+    "./server/control-plane-status": "./src/server/control-plane-status.ts",
     "./server/db": "./src/server/db.ts",
     "./server/env-compat": "./src/server/env-compat.ts",
     "./server/feature-flags": "./src/server/feature-flags.ts",
@@ -31,6 +37,7 @@
     "./server/primitives-memory": "./src/server/primitives-memory.ts",
     "./server/primitives-reconciler": "./src/server/primitives-reconciler.ts",
     "./server/primitives-store": "./src/server/primitives-store.ts",
+    "./server/primitives-watch": "./src/server/primitives-watch.ts",
     "./server/primitives": "./src/server/primitives.ts",
     "./server/primitives-policy": "./src/server/primitives-policy.ts",
     "./server/ready": "./src/server/ready.ts",
@@ -46,6 +53,9 @@
     "./server/agents-controller/mutable-state": "./src/server/agents-controller/mutable-state.ts",
     "./server/agents-controller/rate-limits": "./src/server/agents-controller/rate-limits.ts",
     "./server/agents-controller/*": "./src/server/agents-controller/*.ts",
+    "./server/agentctl-grpc": "./src/server/agentctl-grpc.ts",
+    "./server/agentctl-grpc-config": "./src/server/agentctl-grpc-config.ts",
+    "./server/control-plane-grpc": "./src/server/control-plane-grpc.ts",
     "./server/v1/*": "./src/server/v1/*.ts",
     "./routes/v1/agents": "./src/routes/v1/agents.ts",
     "./routes/v1/agents/$id": "./src/routes/v1/agents/$id.ts",
@@ -59,6 +69,13 @@
     "./routes/v1/orchestration-runs": "./src/routes/v1/orchestration-runs.ts",
     "./routes/v1/orchestration-runs/$id": "./src/routes/v1/orchestration-runs/$id.ts",
     "./routes/v1/runs/$id": "./src/routes/v1/runs/$id.ts",
+    "./routes/api/agents/control-plane/events": "./src/routes/api/agents/control-plane/events.ts",
+    "./routes/api/agents/control-plane/logs": "./src/routes/api/agents/control-plane/logs.ts",
+    "./routes/api/agents/control-plane/resource": "./src/routes/api/agents/control-plane/resource.ts",
+    "./routes/api/agents/control-plane/resources": "./src/routes/api/agents/control-plane/resources.ts",
+    "./routes/api/agents/control-plane/status": "./src/routes/api/agents/control-plane/status.ts",
+    "./routes/api/agents/control-plane/stream": "./src/routes/api/agents/control-plane/stream.ts",
+    "./routes/api/agents/control-plane/summary": "./src/routes/api/agents/control-plane/summary.ts",
     "./runner/*": "./src/runner/*.ts"
   },
   "scripts": {
@@ -70,6 +87,8 @@
   },
   "dependencies": {
     "@bufbuild/cel": "^0.4.0",
+    "@grpc/grpc-js": "^1.14.3",
+    "@grpc/proto-loader": "^0.8.0",
     "@kubernetes/client-node": "^1.4.0",
     "@proompteng/codex": "workspace:*",
     "@proompteng/otel": "workspace:*",
