+++ services/jangar/src/server/__tests__/agents-control-plane-resource.test.ts
@@ -67,7 +67,7 @@ describe('agents control-plane resource route', () => {
 
     const response = await getPrimitiveResource(
       new Request('http://localhost/api/agents/control-plane/resource?kind=Agent&name=agent-a&namespace=agents'),
-      { kubeClient: kube as never },
+      { kubeClient: kube as never, cacheStoreFactory: () => cachedStore as never },
     )
 
     expect(response.status).toBe(200)
@@ -109,7 +109,7 @@ describe('agents control-plane resource route', () => {
 
     const response = await getPrimitiveResource(
       new Request('http://localhost/api/agents/control-plane/resource?kind=Agent&name=agent-a&namespace=agents'),
-      { kubeClient: kube as never },
+      { kubeClient: kube as never, cacheStoreFactory: () => cachedStore as never },
     )
 
     expect(response.status).toBe(200)
@@ -158,7 +158,7 @@ describe('agents control-plane resource route', () => {
 
     const response = await getPrimitiveResource(
       new Request('http://localhost/api/agents/control-plane/resource?kind=Agent&name=agent-a&namespace=agents'),
-      { kubeClient: kube as never },
+      { kubeClient: kube as never, cacheStoreFactory: () => cachedStore as never },
     )
 
     expect(response.status).toBe(200)
